+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to this project will be documented in this file.
 - Support for chacha20-poly1305@openssh.com SSH cipher
 - Quick Settings tile for starting APS in search mode
 
+### Fixed
+
+- Invisible "Copy" icon at the end of field items on password UI
+
 ## [1.15.0] - 2025-07-11
 
 ### Added